Refactor bot services
