+++ bot/services/availability_service.py
