+++ bot/services/database_service.py
