+++ bot/services/event_service.py
