+++ bot/services/user_service.py
