+++ bot/telegram/handlers/inline_handlers.py
@@ -4,13 +4,17 @@ import logging
 
 # Import from config
 from ..config.config import bot
+
+# Import from services
+from services.event_service import getEvent
+from services.database_service import getEntry, setEntry, updateEntry
+from services.event_service import getEventSleepPreferences, join_event
+from services.user_service import updateUsername
+from services.event_service import getUserAvailability, updateUserAvailability
+from services.scheduling_service import calculate_optimal_meeting_time
+
+# Import from utils
 from ..utils.web_app import create_web_app_url
-from ..services.event_service import getEvent
-from ..services.database_service import getEntry, setEntry, updateEntry
-from ..services.event_service import getEventSleepPreferences, join_event
-from ..services.user_service import updateUsername
-from ..services.event_service import getUserAvailability, updateUserAvailability
-from ..services.scheduling_service import calculate_optimal_meeting_time
 from ..utils.message_templates import HELP_MESSAGE
 
 logger = logging.getLogger(__name__)
