+++ bot/telegram/handlers/user_handlers.py
@@ -1,9 +1,15 @@
 import telebot
 from telebot import types
 import logging
-from ..services.scheduling_service import DEFAULT_SLEEP_HOURS
-from ..services.user_service import setUserSleepPreferences
+
+# Import from config
 from ..config.config import bot
+
+# Import from services
+from services.scheduling_service import DEFAULT_SLEEP_HOURS
+from services.user_service import setUserSleepPreferences
+
+# Import from utils
 from ..utils.message_templates import HELP_MESSAGE
 
 logger = logging.getLogger(__name__)
